+++ src/pages/home/Home.jsx
@@ -3,7 +3,7 @@ import Products from "@/components/layout/Products";
 import About from "@/components/global/About";
 import Review from "@/components/layout/Review";
 import Activity from "@/components/layout/Activity";
-import Contact from "@/components/layout/Contact";
+import Footer from "@/components/global/Footer";
 export default function Home(){
   return (
     <div>
@@ -12,7 +12,7 @@ export default function Home(){
       <Products />
       <Review />
       <Activity />
-      <Contact />
+      <Footer />
     </div>
   );
 }
